% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldata-data.R
\docType{data}
\name{alldata}
\alias{alldata}
\title{Enzymes presence/absence matrix}
\format{An object of class \code{"data.frame"}.}
\source{
\href{https://github.com/KARL}
}
\usage{
data(alldata)
}
\description{
The matrix showing the presence (1) or absence (0) of each enzyme
on each genome, and the full lineage for each genome.
}
\examples{
data(alldata)
alldata[1:10,'domain']
}
\references{
Iraola et al. (2016) Unpublished xx:yy-zz.
(\href{https://github.com/KARL})
}
\keyword{alldata}

