% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblR.R
\name{assemblR}
\alias{assemblR}
\title{assemblR}
\usage{
assemblR(r1, r2, mode, out, report)
}
\arguments{
\item{r1}{The filename for fastq containing reads 1.}

\item{r2}{The filename for fastq containing reads 2 (left empty if single reads are assumed).}

\item{mode}{Takes 'fast' for SPAdes running by default (faster), or 'accurate' for SPAdes running with a wider K-mer range and using the '--careful' flag (slower).}

\item{out}{The name of the output file with contigs.}

\item{report}{The name of the assembly report file.}
}
\description{
This function allows to automatically assemble a genome from single or pair-end reads using SPAdes.
}
\examples{
assemblR(r1='reads1.fastq',r2='reads2.fastq',mode='accurate',out='contigs.fasta',report='assembly_report.txt')
}
\keyword{KARL}
\keyword{SPAdes}
\keyword{assembly}

