% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignR.R
\name{assignR}
\alias{assignR}
\title{assignR}
\usage{
assignR(genes, type, qc = 90, id = 50, thrds = 1)
}
\arguments{
\item{genes}{Is the result from findR function, either in nucleotides or amino acids.}

\item{type}{Takes 'prot' or 'nucl' depending on genes parameter datatype.}

\item{qc}{Is the query coverage treshold selected to consider a good enough hit.}

\item{id}{Is the identity treshold selected to consider a good enough hit.}

\item{thrds}{Is the number of threads to be used in Blast searches.}
}
\description{
This function creates a presence/absence vector of enzymes for an input genome.
}
\examples{
p.a.vector<-assignR(genes='pegs.fasta',type='prot',qc=90,id=50,thrds=8)
}
\keyword{KARL}
\keyword{blastp}
\keyword{enzymes}
\keyword{presence/absence}

