% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSVM.R
\name{buildSVM}
\alias{buildSVM}
\title{buildSVM}
\usage{
buildSVM(data = alldata, rank, tax, prefix)
}
\arguments{
\item{data}{The alldata object by default, can be a subset of selected enzymes (tipically by feature selection).}

\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}

\item{tax}{Is the selected taxon to classify at the desired taxonomic rank.}
}
\description{
This function allows to build the classification model for a given taxon using all enzymes or a subset.
}
\examples{
buildSVM(rank='phylum',tax='Cyanobacteria',prefix='Cyanobacteria')
}
\keyword{KARL}
\keyword{RWeka}
\keyword{SMO}
\keyword{SVM}
\keyword{classification}

