% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp1heat.R
\name{comp1heat}
\alias{comp1heat}
\title{comp1heat}
\usage{
comp1heat(rank, tax, lrows = F, reduce = F)
}
\arguments{
\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}

\item{tax}{The selected taxon name.}

\item{lrows}{Takes logical (True or False) to decide plotting rownames (genome names) or not.}

\item{reduce}{Takes logical (True or False) to reduce rows when presence/absence vectors are identical.}
}
\description{
This function produces a heatmap using enzymes presence/absence in a single taxon.
}
\examples{
comp1heat(rank='genus',tax='Salmonella',lrows=FALSE,reduce=TRUE)
comp1heat(rank='genus',tax='Actinobacillus',lrows=T,reduce=F)
}
\keyword{KARL}
\keyword{enzymes}
\keyword{heatmap}
\keyword{presence/absence}

