% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp2heat.R
\name{comp2heat}
\alias{comp2heat}
\title{comp2heat}
\usage{
comp2heat(rank, tax1, tax2, reduce = F)
}
\arguments{
\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}

\item{tax1}{The first taxon name.}

\item{tax2}{The second taxon name.}

\item{reduce}{Takes logical (True or False) to reduce rows when presence/absence vectors are identical.}
}
\description{
This function produces a heatmap using enzymes presence/absence comparing two selected taxa.
}
\examples{
comp2heat(rank='genus',tax1='Actinobacillus',tax2='Bartonella',reduce=FALSE)
}
\keyword{KARL}
\keyword{enzymes}
\keyword{heatmap}
\keyword{presence/absence}

