% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp3heat.R
\name{comp3heat}
\alias{comp3heat}
\title{comp3heat}
\usage{
comp3heat(rank, tax1, tax2, outgroup, reduce = F)
}
\arguments{
\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}

\item{tax1}{The first taxon name.}

\item{tax2}{The second taxon name.}

\item{outgroup}{The third taxon as outgroup.}

\item{reduce}{Takes logical (True or False) to reduce rows when presence/absence vectors are identical.}
}
\description{
This function produces a heatmap using enzymes presence/absence comparing two selected taxa and one selected outgroup.
}
\examples{
comp3heat(rank='genus',tax1='Xylella',tax2='Dyella',outgroup='Nevskia',reduce=FALSE)
}
\keyword{KARL}
\keyword{enzymes}
\keyword{heatmap}
\keyword{outgroup}
\keyword{presence/absence}

