% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compPath.R
\name{compPath}
\alias{compPath}
\title{compPath}
\usage{
compPath(rank, tax1, tax2, pid)
}
\arguments{
\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}

\item{tax1}{The first taxon name.}

\item{tax2}{The second taxon name.}

\item{pid}{Takes a 5-number code for any KEGG Pathway identifier. Can use findPath() function to find the code given a keyword.}
}
\description{
This function calculates the frequency for each enzyme in a pair of given taxa and displays these frequencies over a selected pathway. Outputs a PNG file.
}
\examples{
compPath(rank='class',tax1='Betaproteobacteria',tax2='Mollicutes',pid='00550')
}
\keyword{KARL}
\keyword{KEGG}
\keyword{enzymes}
\keyword{pathways}

