% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consPath.R
\name{consPath}
\alias{consPath}
\title{consPath}
\usage{
consPath(rank, taxa, pid)
}
\arguments{
\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}

\item{taxa}{The taxon name, or a vector of names.}
}
\description{
This function calculates the completeness of each metabolic pathway for a certain taxon.
}
\examples{
taxco<-c('Streptomyces','Acinetobacter','Mycoplasma','Escherichia')
consPath(rank='genus',taxa=taxco)
}
\keyword{KARL}
\keyword{enzymes}
\keyword{pathways}

