% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropUninformative.R
\name{dropUninformative}
\alias{dropUninformative}
\title{dropUninformative}
\usage{
dropUninformative(rank, tax, dwtax = 0.1, uptax = 0.9, cf = 0.9)
}
\arguments{
\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}

\item{tax}{The taxon name.}

\item{dwtax}{The lower frequency cut-off.}

\item{uptax}{The upper frequency cut-off.}

\item{cf}{The correlation cut-off.}
}
\description{
Feature selection by removing uninformative enzymes based on correlation and co-occurrence.
}
\examples{
streptococcus.droped<-dropUninformative(rank='genus',tax='Streptococcus',dwtax=0.1,uptax=0.9,cf=0.9)
}
\keyword{KARL}
\keyword{correlation}
\keyword{enzymes}
\keyword{feature}
\keyword{frequency}
\keyword{selection}

