% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzymesFreq.R
\name{enzymesFreq}
\alias{enzymesFreq}
\title{enzymesFreq}
\usage{
enzymesFreq(rank, tax1, tax2)
}
\arguments{
\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}

\item{tax1}{The name of first taxon.}

\item{tax2}{The name of second taxon.}
}
\description{
This function calculates and plots the frequency of each enzyme in a pair of given taxa.
}
\examples{
enzymesFreq(rank='class',tax='Actinobacteria',tax2='Gammaproteobacteria')
}
\keyword{KARL}
\keyword{enzymes}
\keyword{frequency}

