% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzymesVenn.R
\name{enzymesVenn}
\alias{enzymesVenn}
\title{enzymesVenn}
\usage{
enzymesVenn(elist)
}
\arguments{
\item{elist}{Is the resulting object from getEnzymes().}
}
\description{
This function takes the output from getEnzymes() and draws a Venn diagram.
}
\examples{
enzymesVenn(elist=enzymes)
}
\keyword{KARL}
\keyword{Venn}
\keyword{diagram}
\keyword{enzymes}
\keyword{frequency}

