% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPath.R
\name{findPath}
\alias{findPath}
\title{findPath}
\usage{
findPath(keyword)
}
\arguments{
\item{keyword}{A keyword to link to KEGG.}
}
\description{
This function identifies the 5-number code for KEGG Pathways given a keyword.
}
\examples{
findPath(keyword='peptidoglycan')
}
\keyword{KARL}
\keyword{KEGG}
\keyword{pathway}

