% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findR.R
\name{findR}
\alias{findR}
\title{findR}
\usage{
findR(genome, complete = F, program = "both", outpre)
}
\arguments{
\item{genome}{takes the file name for contigs or completed genome.}

\item{complete}{takes a logical indicating if the genome is closed or not.}

\item{program}{takes 'prodigal', 'glimmer' or 'both' for choosing the prediction algorithm. When 'both' is selected the intersection of individual searches is outputed using blastp.}

\item{outpre}{is the prefix for output files.}
}
\description{
This function performs genes prediction using Prodigal and/or Glimmer.
}
\examples{
pred.genes<-findR(genome='contigs.fasta',complete=F,program='both',outpre='pegs')
}
\keyword{KARL}
\keyword{genes}
\keyword{glimmer}
\keyword{prodigal}

