% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnzymes.R
\name{getEnzymes}
\alias{getEnzymes}
\title{getEnzymes}
\usage{
getEnzymes(rank, tax1, tax2, uptax1 = 0.9, dwtax1 = 0.1, uptax2 = 0.9,
  dwtax2 = 0.1)
}
\arguments{
\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}

\item{tax1}{The first taxon name.}

\item{tax2}{The second taxon name.}

\item{uptax1}{Is the lower frequency cut-off for tax1 (default 0.9).}

\item{dwtax1}{Is the lower frequency cut-off for tax1 (default 0.1).}

\item{uptax2}{Is the lower frequency cut-off for tax2 (default 0.9).}

\item{dwtax2}{Is the lower frequency cut-off for tax2 (default 0.1).}
}
\description{
Outputs a list with taxon-specific, shared and absent enzymes between taxa.
}
\examples{
enzymes<-getEnzymes(rank='class',tax1='Actinobacteria',tax2='Gammaproteobacteria')
}
\keyword{KARL}
\keyword{compare}
\keyword{enzymes}
\keyword{specific}

