% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiR.R
\name{identifiR}
\alias{identifiR}
\title{identifiR}
\usage{
identifiR(inseq, taxout, geneout)
}
\arguments{
\item{inseq}{Input set of genomic contigs or closed genome.}

\item{taxout}{Filename for taxonomic report.}

\item{geneout}{Filename for the resulting 16S sequence found in genome.}
}
\description{
Identifies a genome based on 16S similarity with SILVA database.
}
\examples{
identifiR(inseq='contigs.fasta',taxout='taxreport.txt',geneout='contigs16S.fasta')
}
\keyword{16S}
\keyword{KARL}
\keyword{Rnammer}
\keyword{SILVA}

