% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMAGO.R
\name{iMAGO}
\alias{iMAGO}
\title{iMAGO}
\usage{
iMAGO(vclist, rank)
}
\arguments{
\item{vclist}{Is the output of votingTaxa() function.}

\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}
}
\description{
Calculates MAGO index for re-classification of wrongly classified genomes and outputs membership probabilities.
}
\examples{
actino.imago<-iMAGO(vclist=actino.v7,rank='phylum')
}
\keyword{KARL}
\keyword{MAGO}
\keyword{PCA}
\keyword{classification}
\keyword{reassignment}

