% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeKARL.R
\name{initializeKARL}
\alias{initializeKARL}
\title{initializeKARL}
\usage{
initializeKARL(rank = "all",
  silva.url = "http://www.arb-silva.de/fileadmin/silva_databases/current/Exports/SILVA_123_SSURef_Nr99_tax_silva.fasta.gz",
  min = 2)
}
\arguments{
\item{rank}{takes 'domain', 'phylum', 'class', 'order', 'family', 'genus' or 'all' (default: 'all').}

\item{min}{takes the minimum number of genomes per taxon to build a model (default: 2).}

\item{data}{default is 'alldata' object.}
}
\description{
This function generates and stores models and databases for prediction.
}
\examples{
initializeKARL(rank='all',min=2)
}
\keyword{models}
\keyword{prediction}

