% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koenzymes-data.R
\docType{data}
\name{koenzymes}
\alias{koenzymes}
\title{Links enzymes EC numbers with KEGG Orthology identifiers}
\format{An object of class \code{"character"}.}
\source{
\href{https://github.com/KARL}
}
\usage{
data(koenzymes)
}
\description{
A named vector linking the EC number for each enzyme with
respectives KEGG Orthology identifiers.
}
\examples{
data(koenzymes)
head(koenzymes)
}
\references{
Iraola et al. (2016) Unpublished xx:yy-zz.
(\href{https://github.com/KARL})
}
\keyword{EC}
\keyword{KEGG}
\keyword{Orthology}
\keyword{enzymes}

