% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/li.pathways-data.R
\docType{data}
\name{li.pathways}
\alias{li.pathways}
\title{Links enzymes EC numbers with Pathway identifiers.}
\format{An object of class \code{"list"}.}
\source{
\href{https://github.com/KARL}
}
\usage{
data(li.pathways)
}
\description{
A named list grouping enzymes belonging to each
metabolic pathway following KEGG Pathway identifiers.
}
\examples{
data(li.pathways)
head(li.pathways)
}
\references{
Iraola et al. (2016) Unpublished xx:yy-zz.
(\href{https://github.com/KARL})
}
\keyword{EC}
\keyword{KEGG}
\keyword{enzymes}
\keyword{pathway}

