% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link2path.R
\name{link2path}
\alias{link2path}
\title{link2path}
\usage{
link2path(elist, conf = 0.01, col1 = "darkgreen", col2 = "maroon")
}
\arguments{
\item{elist}{is a list of enzymes resulting from getEnzymes().}

\item{conf}{is a p-value for confidence cut-off.}

\item{col1}{color for first taxon.}

\item{col2}{color for second taxon.}
}
\description{
Identifies and plots significant pathways between pairs of taxa.
}
\examples{
enzymes<-getEnzymes(tax1='Helicobacteraceae',tax2='Enterococcaceae',level='family')
sig.paths<-link2path(elist=enzymes,conf=0.01,col1='blue',col2='maroon')
}
\keyword{KARL}
\keyword{pathway}
\keyword{significance}

