% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onePath.R
\name{onePath}
\alias{onePath}
\title{onePath}
\usage{
onePath(rank, taxa, pid)
}
\arguments{
\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}

\item{taxa}{A vector of names for desired taxa.}
}
\description{
Retrieves the frequency of each enzyme to a certain pathway in a set of taxa.
}
\examples{
tone<-c('Gammaproteobacteria','Mollicutes','Spirochaetia','Actinobacteria')
pone<-onePath(rank='class',taxa=tone,pid='00550')
}
\keyword{Identification}
\keyword{code}
\keyword{desired}
\keyword{for}
\keyword{pathway.}
\keyword{pid}

