\name{...}
\alias{...}
\title{...}
\description{...}
\usage{...}
\arguments{...}
\details{...}
\value{...}
\references{...}
\note{...}
\author{...}
\seealso{...}
\examples{...}
\keyword{...}

# To fill the documentation,
# See link https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Writing-R-documentation-files
