% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcompEC-data.R
\docType{data}
\name{prcompEC}
\alias{prcompEC}
\title{PCA analysis data.}
\format{An object of class \code{"prcomp"}.}
\source{
\href{https://github.com/KARL}
}
\usage{
data(prcompEC)
}
\description{
An object storing a PCA analysis using data contained in
"alldata" object.
}
\examples{
data(prcompEC)
}
\references{
Iraola et al. (2016) Unpublished xx:yy-zz.
(\href{https://github.com/KARL})
}
\keyword{PCA}
\keyword{enzymes}

