% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcompTax.R
\name{prcompTax}
\alias{prcompTax}
\title{prcompTax}
\usage{
prcompTax(rank, taxa, colors, axes = c(1, 2), legpos = "topright")
}
\arguments{
\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}

\item{taxa}{A vector of names for desired taxa.}

\item{colors}{A vector of colors of same length as taxa.}

\item{axes}{A vector of length 2 or 3 indicating which axes to be plotted.}

\item{legpos}{The position of the legend.}
}
\description{
Plots Principal Components in 2D and 3D and allows taxa coloring.
}
\examples{
prcompTax(rank='genus',taxa=c('Streptomyces','Lactobacillus'),axes=c(1,2),colors=c('red','darkgreen'),legpos='topright')
}
\keyword{KARL.}
\keyword{PCA}
\keyword{enzymes}
\keyword{presence/absence}
\keyword{taxa}

