% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictFull.R
\name{predictFull}
\alias{predictFull}
\title{predictFull}
\usage{
predictFull(input)
}
\arguments{
\item{input}{A presence/absence vector or data frame.}
}
\description{
Predicts the whole lineage of a genome or set from domain to genus.
}
\examples{
full.pred<-predictFull(input=inputdata)
}
\keyword{KARL.}
\keyword{enzymes}
\keyword{lineage}
\keyword{predict}
\keyword{presence/absence}

