% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictOne.R
\name{predictOne}
\alias{predictOne}
\title{predictOne}
\usage{
predictOne(rank, tax, input)
}
\arguments{
\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}

\item{tax}{The taxon name.}

\item{input}{A presence/absence vector or data frame.}
}
\description{
Predicts the membership of a genome or set to a particular taxon.
}
\examples{
input.pred<-predictOne(rank='phylum',tax='Proteobacteria',input=inputdata)
}
\keyword{KARL.}
\keyword{enzymes}
\keyword{predict}
\keyword{presence/absence}
\keyword{taxon}

