% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankSlope.R
\name{rankSlope}
\alias{rankSlope}
\title{rankSlope}
\usage{
rankSlope(idata, rank, tax, plotting = T)
}
\arguments{
\item{idata}{Takes the output from dropUninformative() function.}

\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}

\item{tax}{The taxon name. The values of rank and tax must be the same than in the respective dropUninformative() run.}

\item{plotting}{Is a logical indicating if producing plots or not.}
}
\description{
Calculates Information Gain for each enzyme and rank the values using the output from dropUninformative().
}
\examples{
streptococcus.ranked<-rankSlope(idata=streptococcus.droped,rank='genus',tax='Streptococcus',plotting=T)
}
\keyword{Gain}
\keyword{Information}
\keyword{KARL}
\keyword{enzymes}
\keyword{presence/absence}

