% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa-data.R
\docType{data}
\name{taxa}
\alias{taxa}
\title{All taxa.}
\format{An object of class \code{"data.frame"}.}
\source{
\href{https://github.com/KARL}
}
\usage{
data(taxa)
}
\description{
An object storing all genomes with taxonomic data.
}
\examples{
data(taxa)
head(taxa)
}
\references{
Iraola et al. (2016) Unpublished xx:yy-zz.
(\href{https://github.com/KARL})
}
\keyword{genomes}
\keyword{taxa}

