% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxChild.R
\name{taxChild}
\alias{taxChild}
\title{taxChild}
\usage{
taxChild(rank, tax, counting = F)
}
\arguments{
\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}

\item{tax}{The taxon name.}

\item{counting}{Is a logical (T or F) for counting the number of children or not.}
}
\description{
Outputs the children taxa immediately below a certain rank and taxon.
}
\examples{
taxChild(rank='class',tax='Actinobacteria',counting=FALSE)
}
\keyword{KARL}
\keyword{children}
\keyword{taxa}

