% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAll.R
\name{updateAll}
\alias{updateAll}
\title{updateAll}
\usage{
updateAll(mode = "de", replace = T)
}
\arguments{
\item{mode}{Takes 'de' (default) for building the model with the whole data, or 'fs' for building the model after feature selection procedure.}

\item{replace}{Takes a logical (defualt is TRUE) indicating if the original model is replaced by the updated one or not and keeped in the workind directory.}
}
\description{
Updates all classification models without the need of re-installing the package.
}
\examples{
updateAll(mode='de',replace=T)
}
\keyword{KARL}
\keyword{model}
\keyword{update}

