% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateData.R
\name{updateData}
\alias{updateData}
\title{updateData}
\usage{
updateData(route)
}
\arguments{
\item{route}{is the full ftp path where genomes are stored at PATRIC database.}

\item{replace}{takes a logical indicating if replacing old objects by new ones. If FALSE new objects are stored in a 'tmpdir' in the working directory.}
}
\description{
Updates all data used in Explorer and Predictor modules without the need of re-installing the package.
}
\examples{
patric<-'ftp://ftp.patricbrc.org/patric2/patric3/genomes'
updateData(route=patric,replace=F)
}
\keyword{KARL}
\keyword{enzymes}
\keyword{presence/absence}
\keyword{update}

