% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateModel.R
\name{updateModel}
\alias{updateModel}
\title{updateModel}
\usage{
updateModel(rank, tax, mode = "de", replace = T)
}
\arguments{
\item{rank}{Takes 'domain', 'phylum', 'class', 'order', 'family' or 'genus'.}

\item{tax}{The taxon name.}

\item{mode}{Takes 'de' (default) for building the model with the whole data, or 'fs' for building the model after feature selection procedure.}

\item{replace}{Takes a logical (defualt is TRUE) indicating if the original model is replaced by the updated one or not and keeped in the workind directory.}
}
\description{
Updates a selected classification model without the need of re-installing the package.
}
\examples{
updateModel(rank='genus',tax='Escherichia',replace=T)
}
\keyword{KARL}
\keyword{model}
\keyword{update}

